Other people amend and code refactoring
